backend/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
Pillow>=10.0.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.0.0
django-environ
pandas
numpy
scikit-learn
joblib
requests

